chore: Update CORS configuration to allow credentials from client URL
This commit updates the CORS configuration in the `index.js` file to allow credentials from the client URL specified in the `process.env.CLIENT_URL` environment variable. By setting the `origin` to the client URL and enabling `credentials`, the server will now accept requests with cookies from the client.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (`chore` for miscellaneous tasks).