refactor: Add withCredentials option to axios requests
This commit modifies the `getData` and `postData` functions in the `config.ts` file to include the `withCredentials` option in the axios requests. By setting `withCredentials` to `true`, the requests will include cookies, allowing for authenticated requests to the server. This change improves the security and functionality of the API calls.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (`refactor` for code refactoring).