refactor: Remove token logging in getTokenFromCookie function and update cookie settings
This commit refactors the `getTokenFromCookie` function in the `api.ts` file to remove the logging of the token value. Additionally, it updates the cookie settings in both the `getTokenFromCookie` and `checkLocalStorage` functions to remove the unnecessary console.log statements. These changes improve the code cleanliness and remove unnecessary logging.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (`refactor` for code refactoring).