refactor: Update CORS configuration to allow all origins and credentials
This commit updates the CORS configuration in the `index.js` file to allow all origins (`'*'`) and enable credentials. By allowing all origins and enabling credentials, the server will accept requests with cookies from any origin. This change improves the flexibility and functionality of the CORS configuration.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (`refactor` for code refactoring).