refactor: Update CORS configuration to use CLIENT_URL environment variable for origin
This commit updates the CORS configuration in the `index.js` file to use the `process.env.CLIENT_URL` environment variable for the `origin` value. By using the `CLIENT_URL` environment variable, the server will accept requests with cookies from the specified client URL. This change improves the security and flexibility of the CORS configuration.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (`refactor` for code refactoring).